backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy[asyncio]==2.0.35
asyncpg==0.29.0
alembic==1.13.3
pydantic==2.9.2
pydantic-settings==2.5.2
google-auth==2.35.0
google-api-python-client==2.149.0
apscheduler==3.10.4
httpx==0.27.2
python-multipart==0.0.12
